Just added a bit of code to use all available cores.  Setting up for concurrency.
Big question remains: Is it faster/better to block on channel for all agents to send in requests, or is it better for a BufferedChannel to be set up to absorb all requests, then read them out of the BufferedChannel.

The first simplifies the code(?), the second provides better control(?).